psych_app-main/requirements.txt
Flask
Flask-Cors
requests
gunicorn
google-generativeai
